String Views & Constexpr
Switched over to using string views as the primary way of interacting
with rapidxml.  (Yes, won't make this std-lib free, I'm aware of that.)
Second, parsing flags, passed in as template parameters, is now marked
as 'constexpr' on the if-statements, rendering it C++17 or beyond (well,
the string views do that too).